move validate_commit out of voter
